Extracts IntegrationListItem
