+++ pages/integrations/index.tsx
@@ -20,6 +20,7 @@ import { Tooltip } from "@components/Tooltip";
 import ConfirmationDialogContent from "@components/dialog/ConfirmationDialogContent";
 import { FieldsetLegend, Form, InputGroupBox, TextField } from "@components/form/fields";
 import ConnectIntegration from "@components/integrations/ConnectIntegrations";
+import IntegrationListItem from "@components/integrations/IntegrationListItem";
 import { Alert } from "@components/ui/Alert";
 import Badge from "@components/ui/Badge";
 import Button, { ButtonBaseProps } from "@components/ui/Button";
@@ -465,28 +466,6 @@ function ConnectOrDisconnectIntegrationButton(props: {
         </Button>
       )}
     />
-  );
-}
-
-function IntegrationListItem(props: {
-  imageSrc: string;
-  title: string;
-  description: string;
-  actions?: ReactNode;
-  children?: ReactNode;
-}) {
-  return (
-    <ListItem expanded={!!props.children} className={classNames("flex-col")}>
-      <div className={classNames("flex flex-1 space-x-2 w-full p-3 items-center")}>
-        <Image width={40} height={40} src={`/${props.imageSrc}`} alt={props.title} />
-        <div className="flex-grow pl-2 truncate">
-          <ListItemTitle component="h3">{props.title}</ListItemTitle>
-          <ListItemText component="p">{props.description}</ListItemText>
-        </div>
-        <div>{props.actions}</div>
-      </div>
-      {props.children && <div className="w-full border-t border-gray-200">{props.children}</div>}
-    </ListItem>
   );
 }
